Refactor simple value generations
Move the responsibility to generate simple values from
org.javaunit.autoparams package to org.javaunit.autoparams.generator
package.
